src files structure refactoring
