+++ src/components/eMainScreen.tsx
@@ -1,15 +1,12 @@
-import CityCard from '../../components/city-card/city-card';
-
-type MainScreenProps = {
-  placesCount: number;
-};
-
-function MainScreen({ placesCount }: MainScreenProps): JSX.Element {
+// Functional component representing the empty main screen
+function EMainScreen(): JSX.Element {
   return (
     <div className="page page--gray page--main">
+      {/* Header section */}
       <header className="header">
         <div className="container">
           <div className="header__wrapper">
+            {/* Left side of the header */}
             <div className="header__left">
               <a className="header__logo-link header__logo-link--active">
                 <img
@@ -21,8 +18,10 @@ function MainScreen({ placesCount }: MainScreenProps): JSX.Element {
                 />
               </a>
             </div>
+            {/* Navigation section of the header */}
             <nav className="header__nav">
               <ul className="header__nav-list">
+                {/* User profile link */}
                 <li className="header__nav-item user">
                   <a
                     className="header__nav-link header__nav-link--profile"
@@ -35,6 +34,7 @@ function MainScreen({ placesCount }: MainScreenProps): JSX.Element {
                     <span className="header__favorite-count">3</span>
                   </a>
                 </li>
+                {/* Sign out link */}
                 <li className="header__nav-item">
                   <a className="header__nav-link" href="#">
                     <span className="header__signout">Sign out</span>
@@ -46,11 +46,15 @@ function MainScreen({ placesCount }: MainScreenProps): JSX.Element {
         </div>
       </header>
 
-      <main className="page__main page__main--index">
+      {/* Main section */}
+      <main className="page__main page__main--index page__main--index-empty">
+        {/* Heading */}
         <h1 className="visually-hidden">Cities</h1>
+        {/* Tabs */}
         <div className="tabs">
           <section className="locations container">
             <ul className="locations__list tabs__list">
+              {/* List of locations */}
               <li className="locations__item">
                 <a className="locations__item-link tabs__item" href="#">
                   <span>Paris</span>
@@ -67,7 +71,7 @@ function MainScreen({ placesCount }: MainScreenProps): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item tabs__item--active">
+                <a className="locations__item-link tabs__item">
                   <span>Amsterdam</span>
                 </a>
               </li>
@@ -77,55 +81,30 @@ function MainScreen({ placesCount }: MainScreenProps): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="#">
+                <a
+                  className="locations__item-link tabs__item tabs__item--active"
+                  href="#"
+                >
                   <span>Dusseldorf</span>
                 </a>
               </li>
             </ul>
           </section>
         </div>
+        {/* Cities section */}
         <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{`${placesCount} places to stay in Amsterdam`}</b>
-              <form className="places__sorting" action="#" method="get">
-                <span className="places__sorting-caption">Sort by</span>
-                <span className="places__sorting-type" tabIndex={0}>
-                  Popular
-                  <svg className="places__sorting-arrow" width="7" height="4">
-                    <use xlinkHref="#icon-arrow-select"></use>
-                  </svg>
-                </span>
-                <ul className="places__options places__options--custom places__options--opened">
-                  <li
-                    className="places__option places__option--active"
-                    tabIndex={0}
-                  >
-                    Popular
-                  </li>
-                  <li className="places__option" tabIndex={0}>
-                    Price: low to high
-                  </li>
-                  <li className="places__option" tabIndex={0}>
-                    Price: high to low
-                  </li>
-                  <li className="places__option" tabIndex={0}>
-                    Top rated first
-                  </li>
-                </ul>
-              </form>
-              <div className="cities__places-list places__list tabs__content">
-                <CityCard />
-                <CityCard />
-                <CityCard />
-                <CityCard />
-                <CityCard />
+          <div className="cities__places-container cities__places-container--empty container">
+            {/* No places available */}
+            <section className="cities__no-places">
+              <div className="cities__status-wrapper tabs__content">
+                <b className="cities__status">No places to stay available</b>
+                <p className="cities__status-description">
+                  We could not find any property available at the moment in
+                  Dusseldorf
+                </p>
               </div>
             </section>
-            <div className="cities__right-section">
-              <section className="cities__map map"></section>
-            </div>
+            <div className="cities__right-section"></div>
           </div>
         </div>
       </main>
@@ -133,4 +112,4 @@ function MainScreen({ placesCount }: MainScreenProps): JSX.Element {
   );
 }
 
-export default MainScreen;
+export default EMainScreen;
