+++ src/components/mainScreen.tsx
@@ -1,9 +1,19 @@
-function EmptyMainScreen(): JSX.Element {
+import CityCard from './mainCard';
+
+// Define prop types for the MainScreen component
+type MainScreenProps = {
+  placesCount: number;
+};
+
+// MainScreen component
+function MainScreen({ placesCount }: MainScreenProps): JSX.Element {
   return (
     <div className="page page--gray page--main">
+      {/* Header section */}
       <header className="header">
         <div className="container">
           <div className="header__wrapper">
+            {/* Logo */}
             <div className="header__left">
               <a className="header__logo-link header__logo-link--active">
                 <img
@@ -15,8 +25,10 @@ function EmptyMainScreen(): JSX.Element {
                 />
               </a>
             </div>
+            {/* Navigation */}
             <nav className="header__nav">
               <ul className="header__nav-list">
+                {/* User profile */}
                 <li className="header__nav-item user">
                   <a
                     className="header__nav-link header__nav-link--profile"
@@ -29,6 +41,7 @@ function EmptyMainScreen(): JSX.Element {
                     <span className="header__favorite-count">3</span>
                   </a>
                 </li>
+                {/* Sign out */}
                 <li className="header__nav-item">
                   <a className="header__nav-link" href="#">
                     <span className="header__signout">Sign out</span>
@@ -40,11 +53,14 @@ function EmptyMainScreen(): JSX.Element {
         </div>
       </header>
 
-      <main className="page__main page__main--index page__main--index-empty">
+      {/* Main section */}
+      <main className="page__main page__main--index">
+        {/* City tabs */}
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
           <section className="locations container">
             <ul className="locations__list tabs__list">
+              {/* City list items */}
               <li className="locations__item">
                 <a className="locations__item-link tabs__item" href="#">
                   <span>Paris</span>
@@ -61,7 +77,7 @@ function EmptyMainScreen(): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item">
+                <a className="locations__item-link tabs__item tabs__item--active">
                   <span>Amsterdam</span>
                 </a>
               </li>
@@ -71,10 +87,7 @@ function EmptyMainScreen(): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a
-                  className="locations__item-link tabs__item tabs__item--active"
-                  href="#"
-                >
+                <a className="locations__item-link tabs__item" href="#">
                   <span>Dusseldorf</span>
                 </a>
               </li>
@@ -82,17 +95,53 @@ function EmptyMainScreen(): JSX.Element {
           </section>
         </div>
         <div className="cities">
-          <div className="cities__places-container cities__places-container--empty container">
-            <section className="cities__no-places">
-              <div className="cities__status-wrapper tabs__content">
-                <b className="cities__status">No places to stay available</b>
-                <p className="cities__status-description">
-                  We could not find any property available at the moment in
-                  Dusseldorf
-                </p>
+          <div className="cities__places-container container">
+            <section className="cities__places places">
+              {/* Places */}
+              <h2 className="visually-hidden">Places</h2>
+              {/* Display the count of places */}
+              <b className="places__found">{`${placesCount} places to stay in Amsterdam`}</b>
+              {/* Sorting options */}
+              <form className="places__sorting" action="#" method="get">
+                <span className="places__sorting-caption">Sort by</span>
+                <span className="places__sorting-type" tabIndex={0}>
+                  Popular
+                  <svg className="places__sorting-arrow" width="7" height="4">
+                    <use xlinkHref="#icon-arrow-select"></use>
+                  </svg>
+                </span>
+                <ul className="places__options places__options--custom places__options--opened">
+                  <li
+                    className="places__option places__option--active"
+                    tabIndex={0}
+                  >
+                    Popular
+                  </li>
+                  <li className="places__option" tabIndex={0}>
+                    Price: low to high
+                  </li>
+                  <li className="places__option" tabIndex={0}>
+                    Price: high to low
+                  </li>
+                  <li className="places__option" tabIndex={0}>
+                    Top rated first
+                  </li>
+                </ul>
+              </form>
+              {/* List of city cards */}
+              <div className="cities__places-list places__list tabs__content">
+                {/* Render CityCard component */}
+                <CityCard />
+                <CityCard />
+                <CityCard />
+                <CityCard />
+                <CityCard />
               </div>
             </section>
-            <div className="cities__right-section"></div>
+            {/* Right section (map) */}
+            <div className="cities__right-section">
+              <section className="cities__map map"></section>
+            </div>
           </div>
         </div>
       </main>
@@ -100,4 +149,4 @@ function EmptyMainScreen(): JSX.Element {
   );
 }
 
-export default EmptyMainScreen;
+export default MainScreen;
